Moving RegisterCodec utility method from message to transaction package.
