+++ modules/assets/internal/transactions/mint/message.go
@@ -15,7 +15,7 @@ import (
 	"github.com/persistenceOne/persistenceSDK/schema/helpers"
 	"github.com/persistenceOne/persistenceSDK/schema/types"
 	codecUtilities "github.com/persistenceOne/persistenceSDK/utilities/codec"
-	messageUtilities "github.com/persistenceOne/persistenceSDK/utilities/message"
+	"github.com/persistenceOne/persistenceSDK/utilities/transaction"
 )
 
 type message struct {
@@ -41,8 +41,7 @@ func (message message) ValidateBasic() error {
 	return nil
 }
 func (message message) GetSignBytes() []byte {
-	messageCodec := messageUtilities.RegisterCodec(messagePrototype)
-	return sdkTypes.MustSortJSON(messageCodec.MustMarshalJSON(message))
+	return sdkTypes.MustSortJSON(transaction.RegisterCodec(messagePrototype).MustMarshalJSON(message))
 }
 func (message message) GetSigners() []sdkTypes.AccAddress {
 	return []sdkTypes.AccAddress{message.From}
